refactor(watch): isolate file event reconciliation
